requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.2
obs-websocket-py==1.0
httpx==0.27.2
tiktoklive==6.6.5
pyvirtualcam==0.12.0
opencv-python==4.10.0.84
